add page navigation to contests and tasks

+++ bot/handlers/contests_search_handlers.py
@@ -1,6 +1,5 @@
 from aiogram import types
 from aiogram.dispatcher import FSMContext
-from loguru import logger
 
 from loader import dp, bot
 from utils import prepare_task_message
@@ -66,11 +65,14 @@ async def accept_contest_difficulty(query: types.CallbackQuery,
                                    'Не нашлось контестов по заданным параметрам')
             await query.message.delete()
             return
+        
+        await state.update_data(contest=contests)
 
         await bot.send_message(query.from_user.id, 
                                'Выберите один из предложенных контестов',
                                reply_markup=get_all_items_keyboard(contest_data,
-                                                                   contests))    
+                                                                   contests,
+                                                                   current_page=1))    
         await query.message.delete()
         return
     
@@ -84,21 +86,47 @@ async def accept_contest_difficulty(query: types.CallbackQuery,
     
 
 @dp.callback_query_handler(text_contains='contest_data')
-async def get_contest_tasks(query: types.CallbackQuery):
+async def get_contest_tasks(query: types.CallbackQuery,
+                            state: FSMContext):
     query_data = query.data.split(':')
-    contest_id = int(query_data[1])
-    contest = await get_contest(contest_id)
-    await bot.send_message(query.from_user.id, 
-                           'Выберите задание для просмотра подробностей',
-                           reply_markup=get_all_items_keyboard(task_data, 
-                                                               contest.tasks))
-    await query.message.delete()
+    str_page_number = query_data[2]
+    
+    if str_page_number == 'None':
+        contest_id = int(query_data[1])
+        contest = await get_contest(contest_id)
+        
+        await state.update_data(tasks=contest.tasks)
+        
+        await bot.send_message(query.from_user.id, 
+                            'Выберите задание для просмотра подробностей',
+                            reply_markup=get_all_items_keyboard(task_data, 
+                                                                contest.tasks,
+                                                                current_page=1))
+        await query.message.delete()
+        return
+    
+    page_number = int(str_page_number)
+    data = await state.get_data() 
+    await query.message.edit_reply_markup(get_all_items_keyboard(contest_data, 
+                                                                 data['contest'],
+                                                                 current_page=page_number))
 
 
 @dp.callback_query_handler(text_contains='task_data')
-async def get_task_data(query: types.CallbackQuery):
+async def get_task_data(query: types.CallbackQuery,
+                        state: FSMContext):
     query_data = query.data.split(':')
-    task_id = int(query_data[1])
-    task = await get_task(task_id)
-    await bot.send_message(query.from_user.id,
-                           prepare_task_message(task))
+    str_page_number = query_data[2]
+    
+    if str_page_number == 'None':
+        task_id = int(query_data[1])
+        task = await get_task(task_id)
+        await bot.send_message(query.from_user.id,
+                               prepare_task_message(task))
+        return
+
+    page_number = int(str_page_number)
+    data = await state.get_data() 
+    await query.message.edit_reply_markup(get_all_items_keyboard(task_data, 
+                                                                 data['tasks'],
+                                                                 current_page=page_number))
